requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
flask-openapi3==2.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
mysql-connector-python
aniso8601
attrs
click
greenlet
gunicorn
importlib-metadata
itsdangerous
Jinja2
jsonschema
MarkupSafe
nose2
pydantic==1.10.2
pyrsistent
pytz
six
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
typing_extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.1
unidecode
